Fix constrain function.
Constrain was positioning the first note relative to the prior pitch and
then applying tighten() to adjust the remaining pitches. That was
causing the distribution to be heavily weighted toward the low end of
the range, especially in the 'final' etudes.

Also added debug flag to call provide some diagnostic output.